+++ tests/testthat/test-survey.R
@@ -17,8 +17,53 @@ test_that("model_info", {
 })
 
 test_that("find_predictors", {
-  expect_identical(find_predictors(m1), list(conditional = c("ell", "meals", "mobility")))
-  expect_null(find_predictors(m1, effects = "random"))
+  expect_identical(
+    find_predictors(m1),
+    list(conditional = c("ell", "meals", "mobility"), design = c("stype", "fpc"))
+  )
+  expect_identical(
+    find_predictors(m1, component = "conditional"),
+    list(conditional = c("ell", "meals", "mobility"))
+  )
+  expect_identical(
+    find_predictors(m1, component = "design"),
+    list(design = c("stype", "fpc"))
+  )
+  expect_identical(
+    find_predictors(m1, flatten = TRUE),
+    c("ell", "meals", "mobility", "stype", "fpc")
+  )
+  expect_identical(
+    find_predictors(dstrat),
+    list(design = c("stype", "fpc"))
+  )
+  expect_identical(
+    find_predictors(dstrat, flatten = TRUE),
+    c("stype", "fpc")
+  )
+})
+
+test_that("find_variables", {
+  expect_identical(
+    find_variables(m1),
+    list(
+      response = "api00",
+      conditional = c("ell", "meals", "mobility"),
+      design = c("stype", "fpc")
+    )
+  )
+  expect_identical(
+    find_variables(m1, flatten = TRUE),
+    c("api00", "ell", "meals", "mobility", "stype", "fpc")
+  )
+  expect_identical(
+    find_variables(dstrat),
+    list(design = c("stype", "fpc"))
+  )
+  expect_identical(
+    find_variables(dstrat, flatten = TRUE),
+    c("stype", "fpc")
+  )
 })
 
 test_that("find_response", {
@@ -37,7 +82,7 @@ test_that("get_data", {
   expect_equal(nrow(get_data(m1, verbose = FALSE)), 200)
   expect_equal(
     colnames(get_data(m1, verbose = FALSE)),
-    c("api00", "ell", "meals", "mobility", "(weights)")
+    c("stype", "api00", "meals", "ell", "mobility", "pw", "fpc")
   )
 })
 
@@ -68,6 +113,16 @@ test_that("n_obs", {
   expect_equal(n_obs(m1), 200)
 })
 
+test_that("find_weights", {
+  expect_identical(find_weights(m1), "pw")
+  expect_identical(find_weights(dstrat), "pw")
+})
+
+test_that("get_weights", {
+  expect_equal(get_weights(m1), apistrat$pw, tolerance = 1e-4)
+  expect_equal(get_weights(dstrat), apistrat$pw, tolerance = 1e-4)
+})
+
 test_that("linkfun", {
   expect_false(is.null(link_function(m1)))
 })
@@ -93,4 +148,26 @@ test_that("find_parameters", {
 
 test_that("find_statistic", {
   expect_identical(find_statistic(m1), "t-statistic")
+})
+
+test_that("get_data for svyglm and survey.design", {
+  set.seed(123)
+  n <- 5
+  x <- runif(n)
+  y <- runif(n)
+  z <- runif(n)
+  w <- rep(1, n)
+
+  dat <- data.frame(w, x, y, z)
+
+  des <- survey::svydesign(~1, weights = ~w, data = dat)
+  svy_fit <- survey::svyglm(y ~ poly(x, 2), design = des)
+
+  out <- get_data(svy_fit)
+  expect_equal(out$x, dat$x)
+  expect_named(out, c("w", "x", "y"))
+
+  out <- get_data(des)
+  expect_named(out, c("w", "x", "y", "z"))
+  expect_equal(out$x, dat$x)
 })